watcher: arguably slightly better refactor
